Update common-terminology.md
I make a proposal of modified definition. In this definition, I try to put in evidence what is protected, by what, under which circumstances and against whom. The term 'business code and data' is probably inaccurate and could be changed. My point here is to say to mention that the TEE protects the data that is user specific and should be made confidential. My definition is based on a marketing point of view.